removed un-used routes, and moved modal where the graphs are to the dashboard
