--- front-end/src/routes/export/containers/export.js
@@ -1,12 +1,0 @@
-import React,{Component} from 'react';
-
-
-export default class Export extends Component{
-    render(){
-        return (
-              <div>
-              exports
-              </div>
-        );
-    }
-}
