--- front-end/src/routes/export/index.js
@@ -1,2 +1,0 @@
-import Export from './containers/export';
-export default Export;
