--- front-end/src/routes/graphs/index.js
@@ -1,2 +1,0 @@
-import Graphs from './containers/graphs';
-export default Graphs;
