--- front-end/src/routes/startover/containers/startover.js
@@ -1,17 +1,0 @@
-import React,{Component} from 'react';
-import FileUpload from '../../../components/dashboard/Actions/fileUploadActions'
-
-
-
-
-export default class StartOver extends Component{
-
-
-    render(){
-        return (
-          <div>
-     <FileUpload/>
-    </div>
-        );
-    }
-}
