--- front-end/src/routes/startover/index.js
@@ -1,2 +1,0 @@
-import StartOver from './containers/startover';
-export default StartOver;
